Disable polling the messages.
With this on ZAP and the HUD can quickly become unusable.
This is partly because the full messages are returned, including the
requests and responses, while it actually only needs the sort of data
shown in the ZZAP History tab
In any case it looks like its broken - it keeps returning the first 20
messages.
I think it should
* Not poll by default, ie only when the Timeline/History panel is shown
* Use a new API endpoint that just returns the data needed