Add options for enabled, inScopeOnly, devMode and allowUnsafeEval
Fixes #38
Fixes #44

AllowUnsafeEval changes the CSP, and requires development mode.
Switching dev mode on and off generates new events.
All options are persisted, including whether the HUD is on or off.
All options are visible and settable via the API with the exception of
allowUnsafeEval - that cant be set via the API as it feels to dangerous
;)