Alert event performance improvements
The performance issues can be reproduced by spidering bodgeit via the
HUD, ideally with all of the pscan rules installed.
As ZAP processes the pscan queue a large number of alert events are sent
to the HUD (~3500 in my case). The HUD became slower and slower before
often locking up.
The issue turned out to be the number of time we were saving data to
indexdb.
There are 2 parts to this fix:

1. Not saving a tool if the data hasnt changes (an easy win)
1. Processing the alert events in batches - often only 1 or 2 events
will be processed at one time but I've seen 70+ also been handled, which
makes a significant difference.