Browser test API calls
Adds a set of API calls for launching a browser, pointing it at a set of
URLs and checking that the HUD loads.
Has been tested against the Alexa top 200 sites.
Its not perfect - the first one (google) seems to fail to load the HUD
for me even though it works other times.