Only show one flag per alert
To test: spider bodgeit and then run the ascan with just (or at least)
integer overflow.
Then use the HUD - when you add something to the basket a load of alert
flags for integer overflows will be shown.
With this fix just one flag per vuln per field will be shown