Add HUD Errors tool
This is currently shown on the bottom left hand side.
Ideally it will not be included by default when we release, but prior to
that its a good way to highlight and errors that occur :)
Note that this PR changes the std dialog to preserve newlines, which I
think makes formatting easier. The attack tool is the only dialog I
spotted that this cause any problems with...