Initial internationalisation
This is for #23 but it only i18n's one file as I want to see if everyone
agrees with the proposal - ie all of the i18n'd messages at the top.
If thats ok then I'll start on the other files in future PRs.

To test switch ZAP to French, restart it and launch the HUD. The Active
scan tool should be partially translated :)

Its worth noting that \n's (and similar) must be escaped or it breaks
the relevant tool :/