Add option to show/hide side panels - DO NOT MERGE
This is an experiment for feedback.
It moves the 'gear' config icon to the bottom right (where its less
likely to cover important app functionality) and adds a new button with
the HUD icon.
Clicking on the HUD icon shows and hides the side panles, eg in case
they are covering important app functionality.
At the moment the state is not persisted, and ideally the icon would
change colour (eg to grey when the frames are hidden).
I thought this could also be a place for another icon for running hidden
tools (as per #92)
Thoughts ??