Use Gradle to build the add-on
Remove Ant build file and related libraries, the add-on is now built and
deployed with Gradle (as well as running the NPM tasks).
Remove all libraries that are available in Maven Central, that is, all
except the add-ons (Selenium/WebSocket), those were also renamed to have
jar extension so that Eclipse uses them.

Move Java and resources files to default folders `src/main/java/`,
`src/test/java/`, and `src/main/resources/`.
Move bundled add-on files (the ones that are copied to ZAP home dir) to
`src/main/zapHomeFiles/` so that the build picks them for the add-on
manifest (ZapAddOn.xml).
Move UIMessages files to `src/other/resources/UIMessages/`, as with the
bundled add-on files it allows the build to use them to generate the
`i18n.js` file when the add-on is built.
Move JS/webpack related files to `src/test/js/` and `src/webpack/` in an
attempt to organise them. All files were updated accordingly.

The add-on can be built with `gradle assemble` and built/deployed (to
ZAP) with `gradle deployZapAddOn`, the tests are run with `gradle test`.
To build and run the tests (and other additional checks) just use
`gradle build`.

Change the add-on manifest to target ZAP 2.8.0, it requires newer core
API endpoints.

Disable ExampleDotComChromeUnitTest and ExampleDotComFirefoxUnitTest,
its preconditions are not fulfilled by default (ZAP needs to be already
running).

Fix #213 - Migrate build to Gradle