Option changes
Change options so that the HUD can be enables with the ZAP desktop and
daemon mode independently.
Also changed the default enabled options so that ZAP defaults to 'on'
for the desktop 'off' for the daemon, and set all of the current 'dev'
options that are really required for it to run ok.