Load latest upgraded domains
Change the HUD (`serviceworker.js`) to load the upgraded domains from
ZAP after establishing the WebSocket connection to ensure the HUD knows
about them from the beginning, if not the HUD might miss the event that
notifies of the upgrade.
Remove (empty) initialisation of `upgradedDomains` in `utils.js` now
being done in `serviceworker.js`.
Extract a method in `ExtensionHUD` that normalises the upgraded domain.
Add a view to `HudAPI` that exposes the upgraded domains.
Update changelog.

Fix #162 - https upgrade bug on startup
Related to #344 - Make functional tests more reliable