Remove HUD add-on bundled in the downloaded ZAP
Change the `zapDownload` task to delete the HUD add-on bundled in ZAP to
ensure it's used always the local built one (in case of newer version in
the weekly). Also, change the `assembleZapAddOn` task to execute after
the `zapDownload` task to ensure the built add-on has a newer timestamp
(if same version ZAP loads the latest built add-on).