Make tests more reliable
Ensure the HUD is fully initialised before running the actual tests,
otherwise the HUD state might change while the checks are being done
leading to failures.
Ensure frame is restored after each test to have the expected state
in following tests.
Simplify button checks and assertions, with above changes it's no longer
necessary to try restore to the expected frame/state (it already is) and
the number of buttons is now implicitly asserted (either the required
number is found or it fails).
Update ChromeDriver to stable version.
Accept insecure certs when setting up the browsers (not enabled for
Chrome by default, do the same for Firefox for consistency).
Enable output of passed tests when running in Travis CI to not hit the
build timeout limit without output (happening when running the tests
with both browsers).

Part of #344 - Make functional tests more reliable

Signed-off-by: thc202 <thc202@gmail.com>